Proyecto terminado sin BD

+++ QuieroUn10/Controllers/CalendarTasksController.cs
@@ -7,12 +7,15 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using QuieroUn10.Data;
+using QuieroUn10.Filter;
 using QuieroUn10.Models;
 using Rotativa.AspNetCore;
 using Task = QuieroUn10.Models.Task;
 
 namespace QuieroUn10.Controllers
 {
+    [ServiceFilter(typeof(Security))]
+    
     public class CalendarTasksController : Controller
     {
         private readonly QuieroUnDiezDBContex _context;
@@ -23,6 +26,8 @@ namespace QuieroUn10.Controllers
         }
         //https://github.com/esausilva/fullcalendar-aspnet-core/blob/96133f88afd1373c2e34f9f734c7523bf2153117/fullcalendar-core/Scripts/calendar.js
         // GET: CalendarTasks
+
+        [ServiceFilter(typeof(SecurityStudent))]
         public async Task<IActionResult> Index()
         {
             var id = Convert.ToInt32(HttpContext.Session.GetString("user"));
@@ -52,102 +57,81 @@ namespace QuieroUn10.Controllers
 
 
 
-        // GET: CalendarTasks/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
+        /* // GET: CalendarTasks/Details/5
+         public async Task<IActionResult> Details(int? id)
+         {
+             if (id == null)
+             {
+                 return NotFound();
+             }
 
-            var calendarTask = await _context.CalendarTask
-                .Include(c => c.Student)
-                .FirstOrDefaultAsync(m => m.ID == id);
-            if (calendarTask == null)
-            {
-                return NotFound();
-            }
+             var calendarTask = await _context.CalendarTask
+                 .Include(c => c.Student)
+                 .FirstOrDefaultAsync(m => m.ID == id);
+             if (calendarTask == null)
+             {
+                 return NotFound();
+             }
 
-            return View(calendarTask);
-        }
+             return View(calendarTask);
+         }
 
-        // GET: CalendarTasks/Create
-        public IActionResult Create()
-        {
-            ViewData["StudentId"] = new SelectList(_context.Student, "ID", "Name");
-            return View();
-        }
 
-        // POST: CalendarTasks/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,Day,StudentId")] CalendarTask calendarTask)
-        {
-            if (ModelState.IsValid)
-            {
-                _context.Add(calendarTask);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["StudentId"] = new SelectList(_context.Student, "ID", "Name", calendarTask.StudentId);
-            return View(calendarTask);
-        }
 
-        // GET: CalendarTasks/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
+         // GET: CalendarTasks/Edit/5
+         public async Task<IActionResult> Edit(int? id)
+         {
+             if (id == null)
+             {
+                 return NotFound();
+             }
 
-            var calendarTask = await _context.CalendarTask.FindAsync(id);
-            if (calendarTask == null)
-            {
-                return NotFound();
-            }
-            ViewData["StudentId"] = new SelectList(_context.Student, "ID", "Name", calendarTask.StudentId);
-            return View(calendarTask);
-        }
+             var calendarTask = await _context.CalendarTask.FindAsync(id);
+             if (calendarTask == null)
+             {
+                 return NotFound();
+             }
+             ViewData["StudentId"] = new SelectList(_context.Student, "ID", "Name", calendarTask.StudentId);
+             return View(calendarTask);
+         }
 
-        // POST: CalendarTasks/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,Day,StudentId")] CalendarTask calendarTask)
-        {
-            if (id != calendarTask.ID)
-            {
-                return NotFound();
-            }
+         // POST: CalendarTasks/Edit/5
+         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
+         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+         [HttpPost]
+         [ValidateAntiForgeryToken]
+         public async Task<IActionResult> Edit(int id, [Bind("ID,Day,StudentId")] CalendarTask calendarTask)
+         {
+             if (id != calendarTask.ID)
+             {
+                 return NotFound();
+             }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(calendarTask);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!CalendarTaskExists(calendarTask.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["StudentId"] = new SelectList(_context.Student, "ID", "Name", calendarTask.StudentId);
-            return View(calendarTask);
-        }
+             if (ModelState.IsValid)
+             {
+                 try
+                 {
+                     _context.Update(calendarTask);
+                     await _context.SaveChangesAsync();
+                 }
+                 catch (DbUpdateConcurrencyException)
+                 {
+                     if (!CalendarTaskExists(calendarTask.ID))
+                     {
+                         return NotFound();
+                     }
+                     else
+                     {
+                         throw;
+                     }
+                 }
+                 return RedirectToAction(nameof(Index));
+             }
+             ViewData["StudentId"] = new SelectList(_context.Student, "ID", "Name", calendarTask.StudentId);
+             return View(calendarTask);
+         }*/
 
+        [ServiceFilter(typeof(SecurityAdmin))]
         // GET: CalendarTasks/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
@@ -155,27 +139,66 @@ namespace QuieroUn10.Controllers
             {
                 return NotFound();
             }
+            var idC = Convert.ToInt32(HttpContext.Session.GetString("user"));
+            var usuario = _context.UserAccount.Include(r => r.Role).Where(r => r.ID == idC).FirstOrDefault();
 
-            var calendarTask = await _context.CalendarTask
-                .Include(c => c.Student)
-                .FirstOrDefaultAsync(m => m.ID == id);
-            if (calendarTask == null)
+            if (usuario.Role.Name.Equals("STUDENT"))
             {
-                return NotFound();
+                var calendarTask = await _context.CalendarTask
+                    .Include(c => c.Student)
+                    .FirstOrDefaultAsync(m => m.ID == id);
+                if (calendarTask == null)
+                {
+                    return NotFound();
+                }
+                var student = _context.Student.Where(s => s.ID == calendarTask.StudentId).FirstOrDefault();
+                if (student.UserAccountId == usuario.ID)
+                {
+                    return View(calendarTask);
+                }
+                else
+                {
+                    return RedirectToAction("Index", "StudentHasSubjects", new { errorMessage = "No tienes permiso para eliminar tareas del calendario de otros usuarios" });
+                }
+
+            }
+            else
+            {
+                return RedirectToAction("Index", "Subjects", new { errorMessage = "No tiene permiso para eliminar una tarea del calendario" });
             }
 
-            return View(calendarTask);
         }
 
         // POST: CalendarTasks/Delete/5
+        [ServiceFilter(typeof(SecurityAdmin))]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var calendarTask = await _context.CalendarTask.FindAsync(id);
-            _context.CalendarTask.Remove(calendarTask);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var idC = Convert.ToInt32(HttpContext.Session.GetString("user"));
+            var usuario = _context.UserAccount.Include(r => r.Role).Where(r => r.ID == idC).FirstOrDefault();
+
+            if (usuario.Role.Name.Equals("STUDENT"))
+            {
+                var student = _context.Student.Where(s => s.ID == calendarTask.StudentId).FirstOrDefault();
+                if (student.UserAccountId == usuario.ID)
+                {
+                    _context.CalendarTask.Remove(calendarTask);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    return RedirectToAction("Index", "StudentHasSubjects", new { errorMessage = "No tienes permiso para eliminar tareas del calendario de otros usuarios" });
+                }
+
+            }
+            else
+            {
+                return RedirectToAction("Index", "Subjects", new { errorMessage = "No tiene permiso para eliminar una tarea del calendario" });
+            }
+
         }
 
         private bool CalendarTaskExists(int id)
